+++ TSP_MATRIX/TspMatrix.cpp
@@ -20,11 +20,11 @@ TspMatrix::TspMatrix(int n, int **matrix) {
     this->matrix = matrix;
 }
 
-int TspMatrix::calculateCost(int *path) {
+unsigned long long int TspMatrix::calculateCost(int *path) {
     int first = path[0];
     int v1 = path[0];
     int v2;
-    int totalCost = 0;
+    unsigned long long int totalCost = 0;
     for (int i = 1; i < this->n; i++) {
         v2 = path[i];
         totalCost += matrix[v1][v2];
